requirements.txt
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.105.0
graphql-core==3.2.3
greenlet==3.0.2
h11==0.14.0
idna==3.6
pydantic==2.5.2
pydantic-core==2.14.5
python-dateutil==2.8.2
python-multipart==0.0.6
SQLAlchemy==2.0.23
starlette==0.27.0
strawberry-graphql==0.216.1
typing_extensions==4.9.0
uvicorn==0.24.0.post1
